remove type dependency cycle
